feat: deposit token impl
